+++ src/Librarie/Rechizita/Rechizita.hpp
@@ -16,7 +16,7 @@ namespace Librarie{
             Rechizita *operator=(const Rechizita &p);
             ~Rechizita();
             void afiseaza();
-            float pret_redus(){std::cout<<"pret rechizita\n";return pret*0.85;}
+            float pret_redus(){std::cout<<"pret rechizita\n";return (float)(pret*0.85);}
     };
 }
 #endif
